change project names

+++ Chapter 2/PP2.2.c
@@ -3,7 +3,7 @@
 
 int main(void)
 {
-    float rad = 10;
+    int rad = 10;
     float v = 4.0f / 3.0f * M_PI * rad * rad * rad;
 
     printf("The volume is %.3f.\n", v);
